requirements.txt
streamlit>=1.28.0
requests>=2.31.0
qrcode[pil]>=7.4.2
Pillow>=9.5.0